+++ src/index.ts
@@ -61,7 +61,8 @@ export async function codegen(params: ISwaggerOptions) {
   let apiSource = options.useCustomerRequestInstance ? customerServiceHeader(options) : serviceHeader(options)
   // 判断是否是openApi3.0或者swagger3.0
   const isV3 = isOpenApi3(params.openApi || swaggerSource.openapi || swaggerSource.swagger)
-  let requestClasses = Object.entries(requestCodegen(swaggerSource.paths))
+  console.log('isV3', isV3)
+  let requestClasses = Object.entries(requestCodegen(swaggerSource.paths, isV3))
 
   const { models, enums } = isV3
     ? componentsCodegen(swaggerSource.components)
@@ -70,7 +71,7 @@ export async function codegen(params: ISwaggerOptions) {
   // TODO: next next next time
   // if (options.multipleFileMode) {
   if (false) {
-    Object.entries(requestCodegen(swaggerSource.paths)).forEach(([className, requests]) => {
+    Object.entries(requestCodegen(swaggerSource.paths, isV3)).forEach(([className, requests]) => {
       let text = ''
       requests.forEach(req => {
         const reqName = options.methodNameMode == 'operationId' ? req.operationId : req.name